requirements.txt
fastapi
uvicorn[standard]
httpx
SQLAlchemy
databases[postgresql]
alembic
psycopg2-binary
python-dotenv
asyncpg
requests
